sample_codes/cp07_AppRunner/047_AppRunner/docker/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
gradio==5.4.0
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
requests==2.32.3
